Fix blocking while waiting for key
The timeout() for input is set on main window (stdscr), since it seems
that for some ncurses implementations get_wch and getch are the same
(non widechar) wget_wch is used instead. The window passed must the
the main one, the others doesn't have the timeout settings. Another
option is to set wtimeout() on every window.