Properly determine the host of a chess game on initialization
We previously used the length of the invite packet to determine
if we were the host of a chess game. This allowed a friend to
send an invite packet with a length of zero to trick you into
starting a game of chess with yourself as the host. The effect
of this was just a broken game in which neither peer could make
a move.

We now determine who the host is independently and in addition
do more thorough packet size enforcement